refactor(date): move date format to angular pipe
