+++ src/app/core/pipes/date/date.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { CrxDatePipe } from './date.pipe';
+
+@NgModule({
+    declarations: [
+        CrxDatePipe
+    ],
+    exports: [
+        CrxDatePipe
+    ],
+    imports: [
+        CommonModule
+    ]
+})
+export class CrxDateModule {
+}
