+++ src/app/core/pipes/date/index.ts
@@ -1,0 +1,2 @@
+export * from './date.module';
+export * from './date.pipe';
