+++ src/app/features/profile/store/profile.fns.spec.ts
@@ -1,45 +1,7 @@
-import { daySuffix, formatDate, parseUserProfilesFromApiResponse } from '@features/profile/store/profile.fns';
+import { parseUserProfilesFromApiResponse } from '@features/profile/store/profile.fns';
 import { ProfileApiResponse } from '@features/profile/interfaces/profile-api';
 
 describe('Date', () => {
-
-    describe('daySuffix', () => {
-
-        it('should add st suffix', () => {
-
-            expect(daySuffix(1)).toBe('st');
-
-        });
-
-        it('should add nd suffix', () => {
-
-            expect(daySuffix(2)).toBe('nd');
-
-        });
-
-        it('should add rd suffix', () => {
-
-            expect(daySuffix(3)).toBe('rd');
-
-        });
-
-        it('should add th suffix', () => {
-
-            expect(daySuffix(4)).toBe('th');
-
-        });
-
-    });
-
-    describe('formatDate', () => {
-
-        it('should add st suffix', () => {
-
-            expect(formatDate(new Date('1982-08-11T05:19:56.913Z'))).toBe('Aug 11th, 1982');
-
-        });
-
-    });
 
     describe('parseUserProfilesFromApiResponse', () => {
 
@@ -88,7 +50,7 @@ describe('Date', () => {
             const expected = {
                 cellNumber: 'I02 D68-7701',
                 city: 'Selkirk',
-                dateOfBirth: 'Feb 28th, 1955',
+                dateOfBirth: '1955-02-28T10:00:04.795Z',
                 email: 'mia.roy@example.com',
                 firstName: 'Mia',
                 lastName: 'Roy',
